pyproject.toml
[project]
name = "t80m"
version = "1.0.0"
description = "TIC-80 Collection Manager"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "gitpython>=3.1.46",
    "openai>=2.15.0",
    "pillow>=12.1.0",
    "platformdirs>=4.5.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tomli-w>=1.2.0",
    "typer>=0.21.1",
]

[project.scripts]
t80m = "tcolmanager.main:main"

[tool.t80m.initpack]
urls = ["https://pub-a92c0b929eca476cba03da88268d07db.r2.dev/TIC80-collection-2026-01-03.tar.xz"]
sha256 = "5227da792ec8abd05169116dcc082af46c59b0a543dddef7b6e822f4d1601cff"
